Abstracted client behavior into BaseClient
