+++ lib/ethereumex/http_client.ex
@@ -1,22 +1,11 @@
 defmodule Ethereumex.HttpClient do
   use Ethereumex.Client.Macro
+  use Ethereumex.Client.BaseClient  
   import Ethereumex.Config
   @moduledoc false
 
-  @spec single_request(map(), []) :: {:ok, any() | [any()]} | error
-  def single_request(payload, opts \\ []) do
-    payload
-    |> encode_payload
-    |> post_request(opts)
-  end
-
-  @spec encode_payload(map()) :: binary()
-  defp encode_payload(payload) do
-    payload |> Poison.encode!()
-  end
-
   @spec post_request(binary(), []) :: {:ok | :error, any()}
-  defp post_request(payload, opts) do
+  def post_request(payload, opts) do
     headers = [{"Content-Type", "application/json"}]
     options = Ethereumex.Config.http_options()
     url = Keyword.get(opts, :url) || rpc_url()
@@ -44,14 +33,4 @@ defmodule Ethereumex.HttpClient do
     end
   end
 
-  @spec format_batch([map()]) :: [map() | nil | binary()]
-  defp format_batch(list) do
-    list
-    |> Enum.sort(fn %{"id" => id1}, %{"id" => id2} ->
-      id1 <= id2
-    end)
-    |> Enum.map(fn %{"result" => result} ->
-      result
-    end)
-  end
 end
